refactored to remove handlers module, refactored auth code
