+++ .env.sample
@@ -1,6 +1,8 @@
-Secret=
-Client=
-redirect_url=http://localhost:3000/api/callback
+OAUTH_SECRET=
+OAUTH_CLIENT=
+REDIRECT_URL=http://localhost:3000/api/callback
+OAUTH_STATE=
+
 COOKIE_SECRET=
 
 POSTGRES_PASSWORD=
