+++ Dockerfile
@@ -1,0 +1,11 @@
+# Choose whatever you want, version >= 1.16
+FROM golang:1.20-alpine
+
+WORKDIR /app
+
+RUN go install github.com/cosmtrek/air@latest
+
+COPY go.mod go.sum ./
+RUN go mod download
+
+ENTRYPOINT ["air"]
