+++ docker-compose.yml
@@ -1,15 +1,17 @@
 version: '3.9'
 
 services:
-  my-project-with-air:
-    image: cosmtrek/air
-    working_dir: /usr/app
+  api:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    working_dir: /app
     ports:
       - 3000:3000
     env_file:
       - .env.local
     volumes:
-      - ./:/usr/app
+      - ./:/app
     depends_on:
       - db
       - redis
