--- internal/handlers/uploads.go
@@ -1,7 +1,0 @@
-package handlers
-
-import "github.com/gofiber/fiber/v2"
-
-func Upload(c *fiber.Ctx) {
-
-}
