+++ internal/middleware/middleware.go
@@ -4,8 +4,8 @@ import (
 	"eventapi/internal/session"
 	"fmt"
 
-	"github.com/go-redis/redis"
 	"github.com/gofiber/fiber/v2"
+	"github.com/redis/go-redis/v9"
 )
 
 func UseSession(c *fiber.Ctx) error {
@@ -13,16 +13,14 @@ func UseSession(c *fiber.Ctx) error {
 	sId := c.Cookies("session")
 	if sId != "" {
 		sessionData, err := ss.GetSession(c, sId)
+		if err == redis.Nil {
+			ss.ClearSession(c, sId)
+			return c.Redirect("/api/login", 303)
+		}
 
 		if err != nil {
-			if err == redis.Nil {
-				ss.ClearSession(c, sId)
-				return c.Redirect("/api/login", 303)
-			} else {
-				fmt.Println("Could not get session info")
-				return err
-			}
-
+			fmt.Println("Could not get session info")
+			return err
 		}
 
 		//not sure if this is going to work well, thinking i may just hold onto sessionId in context instead
